requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
flake8-builtins==2.1.0
flake8-print==5.0.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
